+++ src/lib/Editor/RemoteResource/AnnotationLoader/parseResponse/responseTypes.js
